pyproject.toml
[project]
name = "content-validation-api"
requires-python = ">=3.11"
version = "0.0.0"
readme = "README.md"
dependencies = [
    "fastapi>=0.115.11",
    "uvicorn==0.21.1",
    "streamlit>=1.44.1",
    "transformers>=4.50.3",
    "torch>=2.6.0",
    "pillow>=11.1.0",
    "types-pillow>=10.2.0.20240822",
    "python-multipart>=0.0.20",
]

[tool.uv]
package = false

[[tool.uv.index]]
name = "kfplc"
url = "https://gitlab.kfplc.com/api/v4/groups/4979/-/packages/pypi/simple"

[project.optional-dependencies]
test = [
    "mypy==1.15.0",
    "pylint==3.3.3",
    "pytest==8.2.0",
    "pytest-cov==4.0.0",
    "pytest-asyncio>=0.25.3",
    "ruff==0.9.6",
    "yamllint>=1.36.0",
    "sqlfluff>=3.3.1",
    "pytest-mock>=3.14.0",
    "types-requests>=2.32.0.20250328",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py311"
exclude = ["docs"]

[tool.ruff.lint]
select = ["F", "E", "W", "I001", "D"]
ignore = ["D203", "D212", "D10"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 80

[tool.pytest.ini_options]
markers = [
    "integration: integration tests (typically slower than unit tests)",
    "system: slower than integration tests and require exernal dependencies"
]

[tool.mypy]
python_version = "3.11"
namespace_packages = true
follow_imports = "silent"
plugins = []
exclude = ["docs"]

[[tool.mypy.overrides]]
module = [
    "uvicorn.*",
    "fastapi_versioning.*",
    "pytest.*",
    "requests.__version__.*",
]
ignore_missing_imports = true

[tool.pylint.main]
disable = [
    "C0114", # missing-module-docstring
    "C0115", # missing-class-docstring
    "C0116", # missing-function-docstring
    "W2301", # unnecessary-ellipsis
    "W0611", # unused-import
    "R0801", # similar-line
    "W0621"  # redefined-outer-name
]
extension-pkg-whitelist = "pydantic"
generated-members = [
    "requests.__version__.*",
    "uvicorn.*",
    "fastapi_versioning.*",
    "pytest.*",
]
jobs = 0
ignore = ["docs",".venv",".cache"]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.design]
max-args = 6

